+++ src/main/java/com/project/emotiondiary/domain/diary/entity/Diary.java
@@ -3,7 +3,7 @@ package com.project.emotiondiary.domain.diary.entity;
 import java.time.LocalDate;
 
 import com.project.emotiondiary.domain.member.entity.Member;
-import com.project.emotiondiary.global.common.entity.BaseTimeEntity;
+import com.project.emotiondiary.global.entity.BaseTimeEntity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
